fixed bug in datastore deletion

+++ src/main/java/org/kuga/kip/datastore/DatastoreController.java
@@ -4,6 +4,10 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 import org.kuga.kip.datastore.field.Field;
 import org.kuga.kip.datastore.field.FieldRepository;
+import org.kuga.kip.datastore.row.Row;
+import org.kuga.kip.datastore.row.RowRepository;
+import org.kuga.kip.datastore.value.ValueRepository;
+import org.kuga.kip.datastore.view.ViewRepository;
 import org.kuga.kip.user.User;
 import org.kuga.kip.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,6 +27,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 
@@ -40,6 +45,18 @@ public class DatastoreController {
 
     @Autowired
     FieldRepository fieldRepository;
+
+    @Autowired
+    RowRepository rowRepository;
+
+    @Autowired
+    FieldValueRepository fieldValueRepository;
+
+    @Autowired
+    ViewRepository viewRepository;
+
+    @Autowired
+    ValueRepository valueRepository;
 
     @PreAuthorize("hasRole('ADMIN')")
     @GetMapping(value = "/")
@@ -91,13 +108,41 @@ public class DatastoreController {
     @DeleteMapping(value = "/delete/{id}")
     public ResponseEntity<HttpStatus> delete(@PathVariable("id") Long id) {
         try {
-            Datastore datastore = datastoreRepository.getById(id);
-            for (User user : datastore.getUsers()) {
-                user.deleteDatastore(datastore);
-                userRepository.save(user);
+            Optional<Datastore> datastoreOptional = datastoreRepository.findById(id);
+            if (datastoreOptional.isPresent()) {
+                Datastore datastore = datastoreOptional.get();
+                for (User user : datastore.getUsers()) {
+                    user.deleteDatastore(datastore);
+                    userRepository.save(user);
+                }
+                Iterator<Row> rowIterator = datastore.getRows().iterator();
+                while (rowIterator.hasNext()) {
+                    Row row = rowIterator.next();
+                    for (FieldValue fieldValue: row.getFieldValues()) {
+                        Field field = fieldValue.getField();
+                        field.deleteFieldValues(fieldValue);
+                        fieldRepository.save(field);
+                        row.deleteFieldValues(fieldValue);
+                        rowRepository.save(row);
+                        fieldValueRepository.delete(fieldValue);
+                    }
+                    rowIterator.remove();
+                    rowRepository.delete(row);
+                }
+                datastoreRepository.save(datastore);
+                Iterator<Field> fieldIterator = datastore.getFields().iterator();
+                while (fieldIterator.hasNext()) {
+                    Field field = fieldIterator.next();
+                    fieldIterator.remove();
+                    fieldRepository.delete(field);
+                }
+                datastoreRepository.save(datastore);
+                datastoreRepository.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
             }
-            datastoreRepository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.OK);
+            else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         }
         catch (Exception ex) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -162,6 +207,7 @@ public class DatastoreController {
                         if (fieldOptional.isPresent()) {
                             Field field = fieldOptional.get();
                             datastore.deleteFields(field);
+                            fieldRepository.delete(field);
                         }
                     }
                 }
